requirements.txt
Flask==3.0.2
Flask-CORS==4.0.0
tensorflow==2.15.0
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1.post1
python-dotenv==1.0.1
pytest==8.0.2
black==24.2.0
isort==5.13.2
flake8==7.0.0